+++ src/UI Elements/BingoLobby.cs
@@ -1,4 +1,6 @@
-﻿using TMPro;
+﻿using System;
+using System.Collections.Generic;
+using TMPro;
 using UltraBINGO.NetworkMessages;
 using UnityEngine;
 using UnityEngine.UI;
@@ -30,8 +32,28 @@ public static class BingoLobby
     public static TMP_Dropdown GameVisibility;
     public static Toggle AllowRejoin;
     public static TMP_Dropdown GameModifiers;
+
+    public static List<GameObject> settingPages;
+    public static GameObject Controls;
+    public static Button Back;
+    public static TextMeshProUGUI PageCount;
+    public static Button Forward;
     
     public static GameObject chatWindow;
+
+    public static int pagePointer = 0;
+
+    public static void SwitchPage(int direction)
+    {
+        pagePointer = Mathf.Clamp(pagePointer + direction, 0, 1);
+
+        PageCount.text = (pagePointer + 1) + "/2";
+
+        foreach (GameObject page in settingPages)
+        {
+            page.SetActive(page.name == ("Page"+(pagePointer+1)));
+        }
+    }
 
     public static void onSettingUpdate(string settingType, int value)
     {
@@ -58,6 +80,22 @@ public static class BingoLobby
             {
                 GameManager.CurrentGame.gameSettingsArray["timeLimit"] = Mathf.Clamp(value,5,30);
                 TimeLimit.text = Mathf.Clamp(value,5,30).ToString();
+                break;
+            }
+            case "gameModifier":
+            {
+                //Force disable the Require P-Rank option if it's enabled due to incompatability
+                if (value > 0)
+                {
+                    GameManager.CurrentGame.gameSettingsArray["requiresPRank"] = 0;
+                    BingoLobby.RequirePRank.isOn = false;
+                    BingoLobby.RequirePRank.interactable = false;
+                }
+                else
+                {
+                    BingoLobby.RequirePRank.interactable = true;
+                }
+
                 break;
             }
             default: break;
@@ -150,42 +188,68 @@ public static class BingoLobby
         
         //Game options
         GameOptions = GetGameObjectChild(BingoLobby,"BingoGameSettings");
+        GameObject pages = GetGameObjectChild(GameOptions, "Pages");
+        settingPages = new List<GameObject>();
+        
+        //Page 1 settings
+        GameObject Page1 = GetGameObjectChild(pages, "Page1");
+        settingPages.Add(Page1);
 
-        MaxPlayers = GetGameObjectChild(GetGameObjectChild(GameOptions,"MaxPlayers"),"Input").GetComponent<TMP_InputField>();
+        MaxPlayers = GetGameObjectChild(GetGameObjectChild(Page1,"MaxPlayers"),"Input").GetComponent<TMP_InputField>();
         MaxPlayers.onEndEdit.AddListener(delegate { onSettingUpdate("maxPlayers",int.Parse(MaxPlayers.text)); });
 
-        MaxTeams = GetGameObjectChild(GetGameObjectChild(GameOptions,"MaxTeams"),"Input").GetComponent<TMP_InputField>();
+        MaxTeams = GetGameObjectChild(GetGameObjectChild(Page1,"MaxTeams"),"Input").GetComponent<TMP_InputField>();
         MaxTeams.onEndEdit.AddListener(delegate { onSettingUpdate("maxTeams",int.Parse(MaxTeams.text)); });
 
-        TimeLimit = GetGameObjectChild(GetGameObjectChild(GameOptions,"TimeLimit"),"Input").GetComponent<TMP_InputField>();
+        TimeLimit = GetGameObjectChild(GetGameObjectChild(Page1,"TimeLimit"),"Input").GetComponent<TMP_InputField>();
         TimeLimit.onEndEdit.AddListener(delegate { onSettingUpdate("timeLimit",int.Parse(TimeLimit.text)); });
 
-        TeamComposition = GetGameObjectChild(GetGameObjectChild(GameOptions,"TeamComposition"),"Dropdown").GetComponent<TMP_Dropdown>();
+        TeamComposition = GetGameObjectChild(GetGameObjectChild(Page1,"TeamComposition"),"Dropdown").GetComponent<TMP_Dropdown>();
         TeamComposition.onValueChanged.AddListener(delegate { onSettingUpdate("teamComposition",TeamComposition.value); });
 
-        Gamemode = GetGameObjectChild(GetGameObjectChild(GameOptions,"Gamemode"),"Dropdown").GetComponent<TMP_Dropdown>();
+        Gamemode = GetGameObjectChild(GetGameObjectChild(Page1,"Gamemode"),"Dropdown").GetComponent<TMP_Dropdown>();
         Gamemode.onValueChanged.AddListener(delegate { onSettingUpdate("gamemode",Gamemode.value); });
 
-        GridSize = GetGameObjectChild(GetGameObjectChild(GameOptions,"GridSize"),"Dropdown").GetComponent<TMP_Dropdown>();
+        GridSize = GetGameObjectChild(GetGameObjectChild(Page1,"GridSize"),"Dropdown").GetComponent<TMP_Dropdown>();
         GridSize.onValueChanged.AddListener(delegate { onSettingUpdate("gridSize",GridSize.value); });
 
-        Difficulty = GetGameObjectChild(GetGameObjectChild(GameOptions,"Difficulty"),"Dropdown").GetComponent<TMP_Dropdown>();
+        Difficulty = GetGameObjectChild(GetGameObjectChild(Page1,"Difficulty"),"Dropdown").GetComponent<TMP_Dropdown>();
         Difficulty.onValueChanged.AddListener(delegate { onSettingUpdate("difficulty",Difficulty.value); });
 
-        RequirePRank = GetGameObjectChild(GetGameObjectChild(GameOptions,"RequirePRank"),"Input").GetComponent<Toggle>();
+        RequirePRank = GetGameObjectChild(GetGameObjectChild(Page1,"RequirePRank"),"Input").GetComponent<Toggle>();
         RequirePRank.onValueChanged.AddListener(delegate { onSettingUpdate("requiresPRank",(RequirePRank.isOn ? 1 : 0)); });
+        
+        //Page 2 settings
+        GameObject Page2 = GetGameObjectChild(pages, "Page2");
+        settingPages.Add(Page2);
 
-        DisableCampaignAltExits = GetGameObjectChild(GetGameObjectChild(GameOptions,"DisableCampaignAltEnds"),"Input").GetComponent<Toggle>();
+        DisableCampaignAltExits = GetGameObjectChild(GetGameObjectChild(Page2,"DisableCampaignAltEnds"),"Input").GetComponent<Toggle>();
         DisableCampaignAltExits.onValueChanged.AddListener(delegate { onSettingUpdate("disableCampaignAltExits",(DisableCampaignAltExits.isOn ? 1 : 0)); });
 
-        GameVisibility = GetGameObjectChild(GetGameObjectChild(GameOptions,"GameVisibility"),"Dropdown").GetComponent<TMP_Dropdown>();
+        GameVisibility = GetGameObjectChild(GetGameObjectChild(Page2,"GameVisibility"),"Dropdown").GetComponent<TMP_Dropdown>();
         GameVisibility.onValueChanged.AddListener(delegate { onSettingUpdate("gameVisibility",GameVisibility.value); });
 
-        AllowRejoin = GetGameObjectChild(GetGameObjectChild(GameOptions,"AllowRejoin"),"Input").GetComponent<Toggle>();
+        AllowRejoin = GetGameObjectChild(GetGameObjectChild(Page2,"AllowRejoin"),"Input").GetComponent<Toggle>();
         AllowRejoin.onValueChanged.AddListener(delegate { onSettingUpdate("allowRejoin",(AllowRejoin.isOn ? 1 : 0)); });
 
-        GameModifiers = GetGameObjectChild(GetGameObjectChild(GameOptions,"GameModifier"),"Dropdown").GetComponent<TMP_Dropdown>();
+        GameModifiers = GetGameObjectChild(GetGameObjectChild(Page2,"GameModifier"),"Dropdown").GetComponent<TMP_Dropdown>();
         GameModifiers.onValueChanged.AddListener(delegate { onSettingUpdate("gameModifier", GameModifiers.value);});
+
+        Controls = GetGameObjectChild(GameOptions, "Controls");
+        Back = GetGameObjectChild(Controls, "Back").GetComponent<Button>();
+        PageCount = GetGameObjectChild(Controls, "PageCount").GetComponent<TextMeshProUGUI>();
+        Forward = GetGameObjectChild(Controls, "Forward").GetComponent<Button>();
+        
+        Back.onClick.AddListener( delegate
+        {
+            SwitchPage(-1);
+        });
+        
+        Forward.onClick.AddListener( delegate
+        {
+            SwitchPage(1);
+        });
+        
 
         if(chatWindow == null)
         {